Move feedsservice to client-lib
